myhistories/pyproject.toml
[tool.poetry]
name = "myhistories"
version = "0.1.0"
description = ""
authors = ["npujol <naivy.luna@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-rest-framework = "^0.1.0"
django-cors-middleware = "^1.5.0"
django-extensions = "^2.2.9"
pyjwt = "^1.7.1"
six = "^1.15.0"
drf-yasg = "^1.17.1"
gtts = "^2.1.1"
celery = "^4.4.5"
django-celery-results = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
